use clap's derive feature for argument parsing
Clap can do most of the argument parsing for us, which allows dropping
lots of parsing code and simplifying the argument-parsing logic.

I tried to keep the same general structure in order to keep this diff
easy to read. In reality, the CompileCommand and InterpreterCommand
should just take an Option<PathBuf> as parameter, and actually reading
the file should be done as part of the command execution -- but that
change would have been quite invasive in an already invasive patch, so
I've left that out of scope for now.